+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 * `[jest-watch]` create new package `jest-watch` to ease custom watch plugin development ([#6318](https://github.com/facebook/jest/pull/6318))
+* Add a config/CLI option `errorOnDeprecated` which makes calling deprecated APIs throw hepful error messages.
 
 ### Fixes